+++ database/migrations/2025_03_02_100050_create_notifications_table.php
@@ -11,7 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('notifications', function (Blueprint $table) {
+        if (!Schema::hasTable('notifications')) {
+            Schema::create('notifications', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('type');
@@ -24,6 +25,7 @@ return new class extends Migration
             $table->boolean('is_sent_push')->default(false);
             $table->timestamps();
         });
+        }
     }
 
     /**
